refactor(lib/date): move date utils to lib directory
