+++ src/components/blog/Hero.astro
@@ -1,5 +1,5 @@
 ---
-import { getFormattedDate } from '@/utils';
+import { getFormattedDate } from '@/lib/date';
 import type { PostCollectionEntry } from '@/types';
 
 interface Props {
