+++ src/components/blog/PostPreview.astro
@@ -1,6 +1,6 @@
 ---
 import type { IElement, PostCollectionEntry } from '@/types';
-import { getFormattedDate } from '@/utils';
+import { getFormattedDate } from '@/lib/date';
 
 interface Props extends IElement {
 	post: PostCollectionEntry;
