+++ src/utils/index.ts
@@ -1,2 +1,1 @@
-export { getFormattedDate } from './date';
 export { elementHasClass, toggleClass } from './domElement';
